+++ lib/ProxyInteractor.dart
@@ -2,21 +2,17 @@ import 'package:flutter/widgets.dart';
 import 'Interactor.dart';
 
 abstract class ProxyInteractorWidget extends StatefulWidget {
-  final WidgetBuilder childBuilder;
+  final Widget child;
 
-  const ProxyInteractorWidget({Key key, @required this.childBuilder})
+  const ProxyInteractorWidget({Key key, @required this.child})
       : super(key: key);
 }
 
 abstract class ProxyInteractor<T extends ProxyInteractorWidget>
     extends Interactor<T> {
   @override
-  Widget build(BuildContext context) => buildWrapper(context, buildChild);
+  Widget build(BuildContext context) => buildWrapper(context, widget.child);
 
   @protected
-  Widget buildWrapper(BuildContext context, WidgetBuilder childBuilder) =>
-      childBuilder(context);
-
-  @protected
-  Widget buildChild(BuildContext context) => widget.childBuilder(context);
+  Widget buildWrapper(BuildContext context, Widget child) => child;
 }